refactor: Moved PaneStrategy to pane module
