+++ leptos-leaflet/src/components/mod.rs
@@ -7,7 +7,7 @@ mod image_overlay;
 mod map_container;
 mod marker;
 mod pane;
-mod pane_strategy;
+
 mod path_options;
 mod polygon;
 mod polyline;
@@ -33,9 +33,9 @@ pub use map_container::{LeafletMap, MapContainer};
 pub use marker::Marker;
 pub use pane::{
     provide_pane_context, provide_pane_context_with_renderer, use_pane_context, Pane, PaneContext,
-    PaneRendererScope,
+    PaneRendererScope, PaneStrategy,
 };
-pub use pane_strategy::PaneStrategy;
+
 pub use path_options::*;
 pub use polygon::Polygon;
 pub use polyline::Polyline;
